examples/MLOps_GitHubActions/AzureFunction/requirements.txt
requests
azure-functions